Boat placement and leagalization checker is working!
Click & drag a new location to place boat and orientation. (can we draw
an arrow under your touch?)
Click and drag the active boat to move it around. 

printBoard() makes sure we know where the boats are.
checkState() maes sure the board is legal. Show continue button
afterward!

Megan contributed to the artwork!